Fixes GH-16 - Live duration updates wrong contraction.
Tags used to find correct duration view were not being reset on completed contractions, causing them to be updated when the top view was not visible (and therefore not the first thing to return from findViewByTag). Now tags are cleared from completed contractions and the view is only updated if it is visible.

Signed-off-by: Ian Lake <ian.hannibal.lake@gmail.com>