Closes GH-19 - Adds 'Keep Screen On' option.
Adds 'Keep Screen On' to the option menu. Unfortunately, Icon menus (such as those that are used pre-Honeycomb), do not support checkable options, hence we use the checkbox yes/no drawables, even if they don't exactly fit in with the other icons.

Note that this seems to have the same issue as seen in GH-18 in that pre-Gingerbread devices need to start/stop the app once without opening the menu for the button to display correctly.

Signed-off-by: Ian Lake <ian.hannibal.lake@gmail.com>