Fixes improper task stack when opening from the Control Widget.
Previously, multiple instances of the MainActivity could be created and
would have separate back stacks associated with them. Also, if the user
was in a dialog/preference page, it would appear when you clicked on the
app launch in the Control Widget. Now, there is only one instance of the
MainActivity and it closes all activities/dialogs on top of it when
opening from the Control Widget.

Signed-off-by: Ian Lake <ian.hannibal.lake@gmail.com>