Adds notification tracking current contraction
Only shows if there is a contraction within the average time frame (as per settings). If the current contraction is ongoing, then the notification is also set to ongoing (i.e., not cancelable) and a 'Stop Contraction' action is available. Otherwise, it is dismissable and shows a 'Start Contraction' action. Uses the chronometer feature to show a count up time of the current duration (for ongoing) or frequency (non-ongoing).