Ensure reset button only disables when there are no contractions at all
Previously, the reset button was disabled if there were no contractions within the current average time span. This unfortunately made it impossible to delete those contractions without starting a new contraction. This splits the Reset menu item and the data it relies on into a separate headless fragment which looks at all available contractions in order to set the enabled/disabled state properly.