Update list empty state with progress spinner pre-load, image+text if no contractions
Previously, the empty state was a single line of unstyled text saying 'Loading Contractions' prior to the first cursor being returned, then nothing after that except the Average bar and the column headers (exactly what isn't helpful).

Now, the prior to first load shows an indeterminate progress spinner (for all of the 1/4 second it is there) before fading into the new empty state: a light gray image and instructions to 'Breathe'. In all cases, the Averages bar and column headers remain hidden until there is at least one contraction - when the first contraction is added, both nicely animate into view.

Note that the image is a placeholder image at this point - a future commit will replace it with the actual image